test just building
