+++ test/com/facebook/buck/swift/SwiftOSXBinaryIntegrationTest.java
@@ -24,7 +24,6 @@ import static org.junit.Assume.assumeThat;
 
 import com.facebook.buck.apple.AppleNativeIntegrationTestUtils;
 import com.facebook.buck.apple.ApplePlatform;
-import com.facebook.buck.apple.AppleTestBuilder;
 import com.facebook.buck.testutil.integration.ProjectWorkspace;
 import com.facebook.buck.testutil.integration.TemporaryPaths;
 import com.facebook.buck.testutil.integration.TestDataHelper;
@@ -33,7 +32,6 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import java.io.IOException;
-import java.nio.file.Paths;
 
 public class SwiftOSXBinaryIntegrationTest {
 
@@ -125,26 +123,5 @@ public class SwiftOSXBinaryIntegrationTest {
     assertThat(
         runResult.getStdout(),
         containsString("Hello ObjC\n"));
-  }
-
-  @Test
-  public void swiftCallingComplexObjCRunsAndPrintsMessageOnOSX() throws IOException {
-    assumeThat(
-        AppleNativeIntegrationTestUtils.isSwiftAvailable(ApplePlatform.MACOSX),
-        is(true));
-    ProjectWorkspace workspace = TestDataHelper.createProjectWorkspaceForScenario(
-        this, "swift_calls_complex_objc", tmp);
-    workspace.setUp();
-
-    workspace.copyRecursively(
-        TestDataHelper.getTestDataDirectory(AppleTestBuilder.class).resolve("fbxctest"),
-        Paths.get("fbxctest"));
-    workspace.writeContentsToPath(
-        "[apple]\n  xctool_path = fbxctest/bin/fbxctest\n",
-        ".buckconfig.local");
-
-    ProjectWorkspace.ProcessResult runResult = workspace.runBuckCommand(
-        "test", ":SwiftCallsComplexObjC#macosx-x86_64");
-    runResult.assertSuccess();
   }
 }
